+++ src/fr/sorbonne_u/components/washingMachine/WashingMachineConnector.java
@@ -74,7 +74,7 @@ public class WashingMachineConnector extends AbstractConnector implements Washin
 	 */
 	
 	@Override
-	public int getCurrentTemperature() throws Exception {
+	public double getCurrentTemperature() throws Exception {
 		return ((WashingMachineCI)this.offering).getCurrentTemperature();
 	}
